feat: wire up stats for sushibar
