+++ src/components/AppBar.tsx
@@ -22,24 +22,23 @@ function AppBar(): JSX.Element {
     const { t } = useTranslation()
     const { pathname } = useLocation()
 
-    const [navClassList, setNavClassList] = useState("w-screen bg-transparent gradiant-border-bottom z-10 backdrop-filter backdrop-blur")
+    const [navClassList, setNavClassList] = useState(
+        'w-screen bg-transparent gradiant-border-bottom z-10 backdrop-filter backdrop-blur'
+    )
 
     const userEthBalance = useETHBalances(account ? [account] : [])?.[account ?? '']
 
     useEffect(() => {
-        if(pathname === '/trade') {
-            setNavClassList("w-screen bg-transparent z-10 backdrop-filter backdrop-blur")
+        if (pathname === '/trade') {
+            setNavClassList('w-screen bg-transparent z-10 backdrop-filter backdrop-blur')
         } else {
-            setNavClassList("w-screen bg-transparent gradiant-border-bottom z-10 backdrop-filter backdrop-blur")
+            setNavClassList('w-screen bg-transparent gradiant-border-bottom z-10 backdrop-filter backdrop-blur')
         }
     }, [pathname])
 
     return (
         <header className="flex flex-row flex-nowrap justify-between w-screen">
-            <Disclosure
-                as="nav"
-                className={navClassList}
-            >
+            <Disclosure as="nav" className={navClassList}>
                 {({ open }) => (
                     <>
                         <div className="px-4 py-1.5">
@@ -71,6 +70,11 @@ function AppBar(): JSX.Element {
                                                     Yield
                                                 </NavLink>
                                             )}
+                                            {chainId === ChainId.MAINNET && (
+                                                <NavLink id={`sushibar-nav-link`} to={'/sushibar'}>
+                                                    SushiBar
+                                                </NavLink>
+                                            )}
                                             {chainId &&
                                                 [ChainId.MAINNET, ChainId.KOVAN, ChainId.BSC, ChainId.MATIC].includes(
                                                     chainId
@@ -79,11 +83,6 @@ function AppBar(): JSX.Element {
                                                         Lend
                                                     </NavLink>
                                                 )}
-                                            {chainId === ChainId.MAINNET && (
-                                                <NavLink id={`stake-nav-link`} to={'/sushi-bar'}>
-                                                    SushiBar
-                                                </NavLink>
-                                            )}
                                             {chainId &&
                                                 [ChainId.MAINNET, ChainId.KOVAN, ChainId.BSC, ChainId.MATIC].includes(
                                                     chainId
