+++ src/pages/SushiBar/InfoCard.tsx
@@ -26,17 +26,20 @@ export default function InfoCard() {
         <div className="flex flex-col max-w-xl w-full mb-2 mt-auto">
             <div className="flex max-w-lg">
                 <div className="text-body font-bold md:text-h5 text-high-emphesis self-end mb-3 md:mb-7">
-                    Maximize yield by staking for xSUSHI
+                    Maximize yield by staking SUSHI for xSUSHI
                 </div>
-                <div className="pl-6 pr-3 mb-1 min-w-max self-start md:hidden">
-                    <img src={XSushiSignSmall} alt="xsushi sign" width="74px" />
-                </div>
+                {/* <div className="pl-6 pr-3 mb-1 min-w-max self-start md:hidden">
+                    <img src={XSushiSignSmall} alt="xsushi sign" />
+                </div> */}
             </div>
-            <div className="text-primary text-xs leading-5 md:text-caption max-w-lg mb-2 md:mb-4 pr-3 md:pr-0">
-                For every swap on Sushi, 0.05% of the swap fee is locked into the liquidity pool awaiting to be served
-                to xSUSHI holders. Anyone can serve it up to reward the swap fees to xSUSHI stakeholders.
+            <div className="text-gray-500 text-sm leading-5 md:text-caption max-w-lg mb-2 md:mb-4 pr-3 md:pr-0">
+                For every swap on the exchange on every chain, 0.05% of the swap fees are distributed as SUSHI
+                proportional to your share of the SushiBar. When your SUSHI is staked into the SushiBar, you recieve
+                xSUSHI in return for voting rights and a fully composable token that can interact with other protocols.
+                Your xSUSHI is continuously compounding, when you unstake you will recive all the SUSHI deposited and
+                additionally accumulated.
             </div>
-            <div className="flex">
+            {/* <div className="flex">
                 <div className="mr-14 md:mr-9">
                     <StyledLink className="text-body whitespace-nowrap text-caption2 md:text-lg md:leading-5">
                         Enter the Kitchen
@@ -47,7 +50,7 @@ export default function InfoCard() {
                         Tips for using xSUSHI
                     </StyledLink>
                 </div>
-            </div>
+            </div> */}
         </div>
     )
 }
