+++ src/pages/SushiBar/TransactionFailedModal.tsx
