+++ src/pages/SushiBar/index.tsx
@@ -1,75 +1,69 @@
 import { Paper } from 'kashi'
 import React from 'react'
 import { useActiveWeb3React } from '../../hooks/useActiveWeb3React'
-import SushiDepositPanel from './SushiDepositPanel'
-import XSushiWithdrawlPanel from './XSushiWithdrawlPanel'
 import { Helmet } from 'react-helmet'
-import { ArrowUpRight } from 'react-feather'
+import XSushiSign from '../../assets/images/xsushi-text-sign.png'
+import InfoCard from './InfoCard'
+import APRCard from './APRCard'
+import StakeCard from './StakeCard'
+import BalanceCard from './BalanceCard'
+import { ChainId } from '@sushiswap/sdk'
+import { SUSHI, XSUSHI } from '../../constants'
+import useTokenBalance from '../../hooks/useTokenBalance'
 
-export default function SushiBar() {
-    const { account } = useActiveWeb3React()
+const mockData = {
+    apr: 21.4,
+    numSushi: 0.1,
+    sushiEarnings: 345.27898,
+    weightedApr: 15.34
+}
+
+export default function XSushi() {
+    const { account, chainId } = useActiveWeb3React()
+
+    const sushiBalance = useTokenBalance(SUSHI[ChainId.MAINNET]?.address ?? '')
+    const xSushiBalance = useTokenBalance(XSUSHI?.address ?? '')
+
     return (
         <>
             <Helmet>
-                <title>Sushi Bar | Sushi</title>
-                <meta
-                    name="description"
-                    content="Exchange SUSHI for xSUSHI to enter the bar and gain a revenue share (0.05%) of income generated by the platform of Sushi products. "
-                />
+                <title>xSUSHI | Sushi</title>
             </Helmet>
-            <div className="flex flex-col max-w-lg w-full mx-auto">
-                <Paper className="bg-dark-800 p-4 space-y-4 mb-4">
-                    <div className="text-lg font-bold text-high-emphesis mb-2">SushiBar: Make SUSHI work for you</div>
-                    <div className="text-sm text-primary">
-                        <p className="mb-4">
-                            Stake your SUSHI into xSUSHI for ~5% APY. No impermanent loss, no loss of governance rights.
-                            Continuously compounding.
-                        </p>
-                        <p>
-                            xSUSHI automatically earn fees (0.05% of all swaps, including multichain swaps) proportional
-                            to your share of the SushiBar.
-                        </p>
+            <div className="flex flex-col w-full">
+                <div className="flex mb-6 justify-center">
+                    <InfoCard />
+                    <div className="hidden md:flex justify-center align-center w-72 ml-6">
+                        <img src={XSushiSign} alt={'xsushi sign'} />
                     </div>
-                    <div className="flex flex-col justify-between my-2 space-y-2 text-sm">
-                        <a
-                            className="text-primary hover:text-high-emphesis flex"
-                            target="_blank"
-                            rel="noreferrer"
-                            href="https://analytics.sushi.com/bar"
-                        >
-                            View SushiBar Stats <ArrowUpRight size="16" />
-                        </a>
-                        {account && (
-                            <a
-                                className="text-primary hover:text-high-emphesis flex"
-                                target="_blank"
-                                rel="noreferrer"
-                                href={'http://analytics.sushi.com/users/' + account}
-                            >
-                                View your SushiBar Portfolio <ArrowUpRight size="16" />
-                            </a>
-                        )}
+                </div>
+                <div className="flex justify-center">
+                    <div className="flex flex-col max-w-xl w-full">
+                        <div className="mb-4">
+                            <APRCard apr={mockData.apr} numSushi={mockData.numSushi} />
+                        </div>
+                        <div>
+                            <StakeCard sushiBalance={sushiBalance} xSushiBalance={xSushiBalance} />
+                        </div>
                     </div>
-                </Paper>
-                <Paper className="bg-dark-900 p-4">
-                    <div className="text-lg font-bold text-high-emphesis mb-4">SUSHI → xSUSHI</div>
-                    <SushiDepositPanel
-                        label={''}
-                        disableCurrencySelect={true}
-                        customBalanceText={'Available to deposit: '}
-                        id="stake-liquidity-token"
-                        buttonText="Deposit"
-                        cornerRadiusBottomNone={true}
-                    />
-                    <XSushiWithdrawlPanel
-                        label={''}
-                        disableCurrencySelect={true}
-                        customBalanceText={'Available to withdraw: '}
-                        id="withdraw-liquidity-token"
-                        buttonText="Withdraw"
-                        cornerRadiusTopNone={true}
-                    />
-                </Paper>
+                    <div className="hidden md:block w-72 ml-6">
+                        <BalanceCard
+                            sushiEarnings={mockData.sushiEarnings}
+                            xSushiBalance={xSushiBalance}
+                            sushiBalance={sushiBalance}
+                            weightedApr={mockData.weightedApr}
+                        />
+                    </div>
+                </div>
+                <div className="flex justify-center w-full">
+                    <div className="md:hidden flex justify-center w-full max-w-xl h-56 mt-6">
+                        <BalanceCard
+                            sushiEarnings={mockData.sushiEarnings}
+                            xSushiBalance={xSushiBalance}
+                            sushiBalance={sushiBalance}
+                            weightedApr={mockData.weightedApr}
+                        />
+                    </div>
+                </div>
             </div>
         </>
     )
