+++ src/routes.tsx
@@ -23,7 +23,6 @@ import RemoveLiquidity from './pages/RemoveLiquidity'
 import { RedirectOldRemoveLiquidityPathStructure } from './pages/RemoveLiquidity/redirects'
 import Saave from './pages/Saave'
 import SushiBar from './pages/SushiBar'
-import XSushi from './pages/XSushi'
 import Trade from './pages/Trade'
 import Swap from './pages/Swap'
 import {
@@ -74,16 +73,14 @@ function Routes(): JSX.Element {
 
             {/* SushiBar Staking */}
             {chainId === ChainId.MAINNET && <Route exact strict path="/sushibar" component={SushiBar} />}
+            {chainId === ChainId.MAINNET && <Route exact strict path="/stake" component={SushiBar} />}
             {/* Tools */}
             {chainId === ChainId.MAINNET && <Route exact strict path="/tools" component={Tools} />}
             {chainId === ChainId.MAINNET && <Route exact strict path="/saave" component={Saave} />}
 
             {/* Pages */}
-            {/* {chainId === ChainId.MAINNET && <Route exact strict path="/trade" component={Trade} />} */}
+            <Route exact strict path="/tradingview" component={Trade} />
             <Route exact strict path="/trade" component={Swap} />
-
-            <Route exact strict path="/sushi-bar" component={SushiBar} />
-            <Route exact strict path="/staking" component={XSushi} />
             <Route exact strict path="/swap" component={Swap} />
             <Route exact strict path="/swap/:outputCurrency" component={RedirectToSwap} />
             <Route exact strict path="/send" component={RedirectPathToSwapOnly} />
